test: split malformed compiled firewall tests
